mcp_server/requirements.txt
azure-functions>=2.0.0
